% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_work_hoursWorkedMainJob.R
\name{adjust_work_hoursWorkedMainJob}
\alias{adjust_work_hoursWorkedMainJob}
\title{Builds a synthetic variable for education attainment}
\usage{
adjust_work_hoursWorkedMainJob(Data)
}
\description{
Builds a synthetic variable for education attainment
}
