% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_age.R
\name{build_demographics_age}
\alias{build_demographics_age}
\title{Builds a synthetic variable for education attainment}
\usage{
build_demographics_age(Data)
}
\description{
Builds a synthetic variable for education attainment
}
