% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_gender.R
\name{build_demographics_gender}
\alias{build_demographics_gender}
\title{Builds a synthetic variable for education attainment}
\usage{
build_demographics_gender(Data)
}
\description{
Builds a synthetic variable for education attainment
}
