% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_race1980standard.R
\name{build_demographics_race1980standard}
\alias{build_demographics_race1980standard}
\title{Builds a synthetic variable for education attainment}
\usage{
build_demographics_race1980standard(Data)
}
\description{
Builds a synthetic variable for education attainment
}
