% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_race2010standard.R
\name{build_demographics_race2010standard}
\alias{build_demographics_race2010standard}
\title{Builds a synthetic variable for education attainment}
\usage{
build_demographics_race2010standard(Data)
}
\description{
Builds a synthetic variable for education attainment
}
