% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_raceDummy.R
\name{build_demographics_raceDummy}
\alias{build_demographics_raceDummy}
\title{Builds a synthetic variable for education attainment}
\usage{
build_demographics_raceDummy(Data)
}
\description{
Builds a synthetic variable for education attainment
}
