% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_attainment.R
\name{build_education_attainment}
\alias{build_education_attainment}
\title{Builds a synthetic variable for education attainment}
\usage{
build_education_attainment(Data)
}
\description{
Builds a synthetic variable for education attainment
}
