% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_levelAttend.R
\name{build_education_levelAttend}
\alias{build_education_levelAttend}
\title{Builds a synthetic variable for education attainment}
\usage{
build_education_levelAttend(Data, ajust_for_age = T)
}
\description{
Builds a synthetic variable for education attainment
}
