% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_literacy.R
\name{build_education_literacy}
\alias{build_education_literacy}
\title{Builds a synthetic variable for education attainment}
\usage{
build_education_literacy(Data)
}
\description{
Builds a synthetic variable for education attainment
}
