% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_yearsSchooling.R
\name{build_education_yearsSchooling}
\alias{build_education_yearsSchooling}
\title{Builds a synthetic variable for education attainment}
\usage{
build_education_yearsSchooling(Data)
}
\description{
Builds a synthetic variable for education attainment
}
