% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_region2010standard.R
\name{build_geography_region2010standard}
\alias{build_geography_region2010standard}
\title{Builds a synthetic variable for education attainment}
\usage{
build_geography_region2010standard(Data)
}
\description{
Builds a synthetic variable for education attainment
}
