% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_regionMCA.R
\name{build_geography_regionMCA}
\alias{build_geography_regionMCA}
\title{Builds a synthetic variable for education attainment}
\usage{
build_geography_regionMCA(Data)
}
\description{
Builds a synthetic variable for education attainment
}
