% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_ruralUrban.R
\name{build_geography_ruralUrban}
\alias{build_geography_ruralUrban}
\title{Builds a synthetic variable for education attainment}
\usage{
build_geography_ruralUrban(Data)
}
\description{
Builds a synthetic variable for education attainment
}
