% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_stateMCA.R
\name{build_geography_stateMCA}
\alias{build_geography_stateMCA}
\title{Builds a synthetic variable for education attainment}
\usage{
build_geography_stateMCA(Data)
}
\description{
Builds a synthetic variable for education attainment
}
