% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_wgt.R
\name{build_identification_wgt}
\alias{build_identification_wgt}
\title{Builds a synthetic variable for education attainment}
\usage{
build_identification_wgt(Data, typeOfweight = "fweight")
}
\description{
Builds a synthetic variable for education attainment
}
