% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year.R
\name{build_identification_year}
\alias{build_identification_year}
\title{Builds a synthetic variable for education attainment}
\usage{
build_identification_year(Data)
}
\description{
Builds a synthetic variable for education attainment
}
