% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_income_incomeOtherSources.R
\name{build_income_incomeOtherSources}
\alias{build_income_incomeOtherSources}
\title{Builds a synthetic variable for education attainment}
\usage{
build_income_incomeOtherSources(Data, deflate_values = T)
}
\description{
Builds a synthetic variable for education attainment
}
