% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_classWorker.R
\name{build_work_classWorker}
\alias{build_work_classWorker}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_classWorker(Data)
}
\description{
Builds a synthetic variable for education attainment
}
