% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_econActivity.R
\name{build_work_econActivity}
\alias{build_work_econActivity}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_econActivity(Data)
}
\description{
Builds a synthetic variable for education attainment
}
