% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_employRecordBook.R
\name{build_work_employRecordBook}
\alias{build_work_employRecordBook}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_employRecordBook(Data)
}
\description{
Builds a synthetic variable for education attainment
}
