% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_hoursWorkedAllJobs.R
\name{build_work_hoursWorkedAllJobs}
\alias{build_work_hoursWorkedAllJobs}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_hoursWorkedAllJobs(Data)
}
\description{
Builds a synthetic variable for education attainment
}
