% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_hoursWorkedMainJob.R
\name{build_work_hoursWorkedMainJob}
\alias{build_work_hoursWorkedMainJob}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_hoursWorkedMainJob(Data)
}
\description{
Builds a synthetic variable for education attainment
}
