% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_socialSecurityContrib.R
\name{build_work_socialSecurityContrib}
\alias{build_work_socialSecurityContrib}
\title{Builds a synthetic variable for education attainment}
\usage{
build_work_socialSecurityContrib(Data)
}
\description{
Builds a synthetic variable for education attainment
}
