% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_demographics.R
\name{harmonize_demographics}
\alias{harmonize_demographics}
\title{Builds a synthetic variable for education attainment}
\usage{
harmonize_demographics(Data)
}
\description{
Builds a synthetic variable for education attainment
}
