% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_education.R
\name{harmonize_education}
\alias{harmonize_education}
\title{Builds a synthetic variable for education attainment}
\usage{
harmonize_education(Data)
}
\description{
Builds a synthetic variable for education attainment
}
