% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_geography.R
\name{harmonize_geography}
\alias{harmonize_geography}
\title{Builds a synthetic variable for education attainment}
\usage{
harmonize_geography(Data)
}
\description{
Builds a synthetic variable for education attainment
}
