% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_identification.R
\name{harmonize_identification}
\alias{harmonize_identification}
\title{Builds a synthetic variable for education attainment}
\usage{
harmonize_identification(Data)
}
\description{
Builds a synthetic variable for education attainment
}
