% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_work.R
\name{harmonize_work}
\alias{harmonize_work}
\title{Builds a synthetic variable for education attainment}
\usage{
harmonize_work(Data)
}
\description{
Builds a synthetic variable for education attainment
}
